Master collection created automatically, no matter what. "Enable multiple collections" is now optional

+++ src/Api/Resource/EventTagsResource.php
@@ -12,11 +12,21 @@
 
         /**
          * Create a new tag
-         * @todo: permissions
+         * @throws ApiException
          * @return void
          */
         protected function httpPost(){
+            // Permission check
+            if( ! $this->getGenericTaskPermissionObj()->canCreateTag() ){
+                throw ApiException::permissionInvalid("You do not have permission to create tags.");
+            }
+            // Get passed data
             $data = $this->scrubbedPostData();
+            // Is displayText property set?
+            if( empty($data->displayText) ){
+                throw ApiException::generic("displayText property must be set");
+            }
+            // All good
             $tagObj = EventTag::createOrGetExisting($data);
             $this->setResponseData($tagObj);
             $this->setResponseCode(Response::HTTP_CREATED);
@@ -24,19 +34,37 @@
 
         /**
          * @param $id
+         * @throws ApiException
+         * @return void
          */
         protected function httpPut( $id ){
+            // Permission check
+            if( ! $this->getGenericTaskPermissionObj()->canCreateTag() ){
+                throw ApiException::permissionInvalid("You do not have permission to create tags.");
+            }
+            // Get passed data
+            $data = $this->scrubbedPostData();
+            // Is displayText property set?
+            if( empty($data->displayText) ){
+                throw ApiException::generic("displayText property must be set");
+            }
+
             $tagObj = EventTag::getByID($id);
-            $tagObj->update($this->scrubbedPostData());
+            $tagObj->update($data);
             $this->setResponseData($tagObj);
             $this->setResponseCode(Response::HTTP_ACCEPTED);
         }
 
         /**
          * @param $id
-         * @todo: permissions, error handling
+         * @throws ApiException
+         * @return void
          */
         protected function httpDelete( $id ){
+            // Permission check
+            if( ! $this->getGenericTaskPermissionObj()->canCreateTag() ){
+                throw ApiException::permissionInvalid("You do not have permission to create tags.");
+            }
             EventTag::getByID($id)->delete();
         }
